:truck: [Utilities] Rename dom to detect
