+++ src/utilities/detect.ts
